requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
sqlalchemy>=2.0.0
pytest>=8.0.0
pytest-mock>=3.12.0
requests>=2.31.0
pytest-cov>=4.1.0
mutmut>=2.4.4
hypothesis>=6.98.0
locust>=2.24.0
bandit>=1.7.7
httpx>=0.27.0
pymongo>=4.15.5
pandas>=2.3.3
tabulate>=0.9.0
matplotlib>=3.10.8
pytest-assume>=2.4.0
psycopg2-binary>=2.9.9
